snowflake_demo/requirements.txt
snowflake-connector-python==3.6.0
pandas>=2.2.0
python-dotenv==1.0.0
numpy>=1.26.0
